Add ASR pre-commit hook with extra wrapper
The ASR pre-commit hook that came from Hello Dotty calls a `docker
compose` command to run linting. This is added to all projects, those
that are Dockerized and those that are not. It throws an error for the
ones that do not have docker compose files. This adds a check for a
docker-compose file in the project before calling the linting check via
compose.