Add Chat GPT helper methods
Add a gptcp for getting the contents of a single file in a copy and
paste block wrapped with tags that show the begin/end of the file
contents.

Add a gptcpdir that gets all files in a directory up to a max byte size.
Similar to the single helper but is handing for multiple files in a
directory when providing context to a chat.